Added a guard to previous name details
This is to resolve an issue where data is being transmitted from the frontend with no legal name, but is being forwarded on in error as if a name has been supplied.

The actual name is blank, but the data suggests that there should be a name present. We suspect this is currently breaking API 5 calls into ETMP.