AMLS-3766 Altered logic for when Supervision returns 'no'
Now, the logic to determine whether all of the Supervision fields are pre-filled with 'no' looks at whether or not the activities list includes ASP and/or TCSP. If not, the None is returned.

This is because the data supplier will return an empty node for `aspOrTcsp` if the user has filled in 'no' to all of those questions on the frontend.

This commit also contains some re-writing of the Supervision spec, which takes a property-based testing approach.