Defaulted the 'dateChangeFlag'
This is causing an issue with acceptance tests when the data is being overridden by another file from stubs. This value is normally updated based on the line id's for trading premises, but in the case where an overriding JSON file is used, the line ids usually don't match up, so the value is never set.