DDCYLS-2315
We've noticed some dates being stored like this in the database:
2022-03-03 10:10:10.123Z
rather than the regular format, this is causing major issues so a custom date time
format has been written for the created date field of the Payment model.

However this needs to be investigated further when we don't have as many live issues
it may require some data migration in the future. We should ideally stick for the ISO 8601
format